Reorganize jina v1 embeddings
